+++ CHANGELOG.md
@@ -22,3 +22,4 @@
 * Set `answer.option` to default option for `drop_down` and `option_list` questions in `Order.construct`
 * Set `answer.value` to `min` for `spin_button` questions in `Order.construct`
 * Add afterRequest callback to transform `product.attendeeQuestions` to `ActiveResource.Collection` with camelized values
+* Add order afterRequest for matching attendee collection size to order quantity